+++ app/components/UI/Navbar/index.js
@@ -13,7 +13,7 @@ import MaterialCommunityIcon from 'react-native-vector-icons/MaterialCommunityIc
 import URL from 'url-parse';
 import { strings } from '../../../../locales/i18n';
 import AppConstants from '../../../core/AppConstants';
-import TabCountIcon from '../../Views/BrowserTab/Tabs/TabCountIcon';
+import TabCountIcon from '../../UI/Tabs/TabCountIcon';
 const HOMEPAGE_URL = 'about:blank';
 
 const styles = StyleSheet.create({
