+++ app/components/UI/Tabs/TabCountIcon/index.test.js
