+++ app/components/UI/Tabs/TabThumbnail/index.test.js
